[engine] ui anchoring main part done

+++ PHEngine/PHEngine/src/Core/GameCore/GUI/UiElements/Transform2D/BoundingBox2D.cpp
@@ -6,12 +6,14 @@ namespace EngineCore
    {
 
       BoundingBox2D::BoundingBox2D()
-          : mOrigin(), mHalfExtent()
+          : mOrigin(),
+            mHalfExtent()
       {
       }
 
-      BoundingBox2D::BoundingBox2D(const glm::vec2 &origin, const glm::vec2 &halfExtent)
-          : mOrigin(origin), mHalfExtent(halfExtent)
+      BoundingBox2D::BoundingBox2D(const glm::ivec2 &origin, const glm::ivec2 &halfExtent)
+          : mOrigin(origin),
+            mHalfExtent(halfExtent)
       {
       }
 
@@ -19,22 +21,22 @@ namespace EngineCore
       {
       }
 
-      glm::vec2 BoundingBox2D::GetOrigin() const
+      glm::ivec2 BoundingBox2D::GetOrigin() const
       {
          return mOrigin;
       }
 
-      glm::vec2 BoundingBox2D::GetHalfExtent() const
+      glm::ivec2 BoundingBox2D::GetHalfExtent() const
       {
          return mHalfExtent;
       }
 
-      glm::vec2 BoundingBox2D::GetMax() const
+      glm::ivec2 BoundingBox2D::GetMax() const
       {
          return mOrigin + mHalfExtent;
       }
 
-      glm::vec2 BoundingBox2D::GetMin() const
+      glm::ivec2 BoundingBox2D::GetMin() const
       {
          return mOrigin - mHalfExtent;
       }
